requirements.txt
# requirements.txt
tensorflow
scikit-learn
pandas
matplotlib
seaborn
fastapi
uvicorn[standard]
python-multipart
pillow
transformers
torch
pytest